+++ packages/omni_mapper_generator/test/src/generator_test_inputs.dart
@@ -460,6 +460,37 @@ class ModelL {
   ModelL({required this.userId, required this.title});
 }
 
+// --- FEATURE 1: Multiple Sources ---
+class SourceX {
+  final int id;
+  SourceX({required this.id});
+}
+
+class SourceY {
+  final String name;
+  SourceY({required this.name});
+}
+
+class TargetMultiple {
+  final int id;
+  final String name;
+  TargetMultiple({required this.id, required this.name});
+}
+
+@ShouldGenerate(r'''
+class MultipleSourcesMapperImpl extends MultipleSourcesMapper {
+  @override
+  TargetMultiple toTarget(SourceX x, SourceY y) {
+    final target = TargetMultiple(id: x.id, name: y.name);
+    return target;
+  }
+}
+''')
+@OmniMapper()
+abstract class MultipleSourcesMapper {
+  TargetMultiple toTarget(SourceX x, SourceY y);
+}
+
 // --- FEATURE 7: Factory Constructors (Freezed) ---
 class FreezedLikeModel {
   final String title;